[api] Start support for GAN
