+++ api/src/main/java/ai/djl/modality/cv/input/ImageNetCategory.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * Copyright 2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance
  * with the License. A copy of the License is located at
@@ -10,7 +10,7 @@
  * OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions
  * and limitations under the License.
  */
-package ai.djl.examples.inference.biggan;
+package ai.djl.modality.cv.input;
 
 import java.io.IOException;
 import java.nio.file.Files;
@@ -22,11 +22,12 @@ import java.util.concurrent.ConcurrentHashMap;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public final class BigGANCategory {
-    private static final Logger logger = LoggerFactory.getLogger(BigGANCategory.class);
+/** A container for an ImageNet class. */
+public final class ImageNetCategory {
+    private static final Logger logger = LoggerFactory.getLogger(ImageNetCategory.class);
 
     public static final int NUMBER_OF_CATEGORIES = 1000;
-    private static final Map<String, BigGANCategory> CATEGORIES_BY_NAME =
+    private static final Map<String, ImageNetCategory> CATEGORIES_BY_NAME =
             new ConcurrentHashMap<>(NUMBER_OF_CATEGORIES);
     private static String[] categoriesById;
 
@@ -37,25 +38,41 @@ public final class BigGANCategory {
         try {
             parseCategories();
         } catch (IOException e) {
-            logger.error("Error parsing the ImageNet categories: {}", e);
+            logger.error("Error parsing the ImageNet categories:", e);
         }
         createCategoriesByName();
     }
 
-    private BigGANCategory(int id, String[] names) {
+    private ImageNetCategory(int id, String[] names) {
         this.id = id;
         this.names = names;
     }
 
+    /**
+     * Returns the category ID.
+     *
+     * @return the category ID
+     */
     public int getId() {
         return id;
     }
 
+    /**
+     * Returns the list of different names for this category.
+     *
+     * @return the list of different names for this category
+     */
     public String[] getNames() {
         return names.clone();
     }
 
-    public static BigGANCategory id(int id) {
+    /**
+     * Get the ImageNet category with the ID.
+     *
+     * @param id of the category
+     * @return the corresponding {@link ImageNetCategory}
+     */
+    public static ImageNetCategory id(int id) {
         String names = categoriesById[id];
         int index = names.indexOf(',');
         if (index < 0) {
@@ -65,17 +82,24 @@ public final class BigGANCategory {
         }
     }
 
-    public static BigGANCategory of(String name) {
+    /**
+     * Get the ImageNet category with one of its names.
+     *
+     * @param name of the category
+     * @return the corresponding {@link ImageNetCategory}
+     */
+    public static ImageNetCategory of(String name) {
         if (!CATEGORIES_BY_NAME.containsKey(name)) {
             throw new IllegalArgumentException(name + " is not a valid category.");
         }
         return CATEGORIES_BY_NAME.get(name);
     }
 
+    /** Create structure to map from ImageNet label to the corresponding ImageNetCategory object. */
     private static void createCategoriesByName() {
         for (int i = 0; i < NUMBER_OF_CATEGORIES; i++) {
             String[] categoryNames = categoriesById[i].split(", ");
-            BigGANCategory category = new BigGANCategory(i, categoryNames);
+            ImageNetCategory category = new ImageNetCategory(i, categoryNames);
 
             for (String name : categoryNames) {
                 CATEGORIES_BY_NAME.put(name, category);
@@ -83,14 +107,21 @@ public final class BigGANCategory {
         }
     }
 
+    /**
+     * Parse all the ImageNet labels from a text file and store them in an array where the index is
+     * the ID.
+     *
+     * @throws IOException if an I/O error occurs reading from the file or a malformed or unmappable
+     *     byte sequence is read
+     */
     private static void parseCategories() throws IOException {
-        String filePath = "src/main/resources/categories.txt";
+        String filePath =
+                "../model-zoo/src/test/resources/mlrepo/model/cv/image_classification/ai/djl/zoo/synset_imagenet.txt";
 
         List<String> fileLines = Files.readAllLines(Paths.get(filePath));
         List<String> categories = new ArrayList<>(NUMBER_OF_CATEGORIES);
         for (String line : fileLines) {
-            int nameIndex = line.indexOf(':') + 2;
-            categories.add(line.substring(nameIndex));
+            categories.add(line.substring(10));
         }
 
         categoriesById = categories.toArray(new String[] {});
