+++ api/src/main/java/ai/djl/modality/cv/input/package-info.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * Copyright 2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance
  * with the License. A copy of the License is located at
@@ -11,5 +11,5 @@
  * and limitations under the License.
  */
 
-/** Contains an examples using BigGAN from DeepMind (Google). */
-package ai.djl.examples.inference.biggan;
+/** Contains input types used in various computer vision applications. */
+package ai.djl.modality.cv.input;
